+++ app/src/main/java/droidkaigi/github/io/challenge2019/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package droidkaigi.github.io.challenge2019
+package droidkaigi.github.io.challenge2019.ui.main
 
 import android.annotation.SuppressLint
 import android.app.Activity
@@ -11,15 +11,17 @@ import android.os.Bundle
 import android.support.v4.widget.SwipeRefreshLayout
 import android.support.v7.widget.DividerItemDecoration
 import android.support.v7.widget.RecyclerView
-import android.util.Log
 import android.view.MenuItem
 import android.view.View
 import android.widget.ProgressBar
 import com.squareup.moshi.Types
+import droidkaigi.github.io.challenge2019.*
 import droidkaigi.github.io.challenge2019.data.api.HackerNewsApi
 import droidkaigi.github.io.challenge2019.data.api.response.Item
 import droidkaigi.github.io.challenge2019.data.db.ArticlePreferences
 import droidkaigi.github.io.challenge2019.data.db.ArticlePreferences.Companion.saveArticleIds
+import droidkaigi.github.io.challenge2019.ui.story.StoryActivity
+import droidkaigi.github.io.challenge2019.ui.story.StoryAdapter
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -65,9 +67,10 @@ class MainActivity : BaseActivity() {
             stories = mutableListOf(),
             onClickItem = { item ->
                 val itemJson = itemJsonAdapter.toJson(item)
-                val intent = Intent(this@MainActivity, StoryActivity::class.java).apply {
-                    putExtra(StoryActivity.EXTRA_ITEM_JSON, itemJson)
-                }
+                val intent =
+                    Intent(this@MainActivity, StoryActivity::class.java).apply {
+                        putExtra(StoryActivity.EXTRA_ITEM_JSON, itemJson)
+                    }
                 startActivityForResult(intent)
             },
             onClickMenuItem = { item, menuItemId ->
@@ -81,11 +84,12 @@ class MainActivity : BaseActivity() {
                             override fun onResponse(call: Call<Item>, response: Response<Item>) {
                                 response.body()?.let { newItem ->
                                     val index = storyAdapter.stories.indexOf(item)
-                                    if (index == -1 ) return
+                                    if (index == -1) return
 
                                     storyAdapter.stories[index] = newItem
                                     runOnUiThread {
-                                        storyAdapter.alreadyReadStories = ArticlePreferences.getArticleIds(this@MainActivity)
+                                        storyAdapter.alreadyReadStories =
+                                            ArticlePreferences.getArticleIds(this@MainActivity)
                                         storyAdapter.notifyItemChanged(index)
                                     }
                                 }
