+++ app/src/main/java/droidkaigi/github/io/challenge2019/ui/story/StoryActivity.kt
@@ -1,4 +1,4 @@
-package droidkaigi.github.io.challenge2019
+package droidkaigi.github.io.challenge2019.ui.story
 
 import android.annotation.SuppressLint
 import android.app.Activity
@@ -15,6 +15,9 @@ import android.webkit.WebView
 import android.webkit.WebViewClient
 import android.widget.ProgressBar
 import com.squareup.moshi.Types
+import droidkaigi.github.io.challenge2019.BaseActivity
+import droidkaigi.github.io.challenge2019.R
+import droidkaigi.github.io.challenge2019.Util
 import droidkaigi.github.io.challenge2019.data.api.HackerNewsApi
 import droidkaigi.github.io.challenge2019.data.api.response.Item
 import retrofit2.Call
